Handle `@typechecked` and `@overload` properly
